[INJIMOB-891] refactor changes
