+++ android/app/src/main/java/io/mosip/residentapp/InjiPackage.java
@@ -22,6 +22,7 @@ public class InjiPackage implements ReactPackage {
         modules.add(new InjiVciClientModule(reactApplicationContext));
         modules.add(new RNPixelpassModule(reactApplicationContext));
         modules.add(new RNSecureKeystoreModule(reactApplicationContext));
+        modules.add(new RNVersionModule());
         modules.add(new RNWalletModule(new RNEventEmitter(reactApplicationContext), new Wallet(reactApplicationContext), reactApplicationContext));
         modules.add(new RNVerifierModule(new RNEventEmitter(reactApplicationContext), new Verifier(reactApplicationContext), reactApplicationContext));
         return modules;
