+++ ios/Inji.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,5 +1,5 @@
 {
-  "originHash" : "a6543a04f734b9c7569cf659b2ca72978ddd4cb37e1d3cb84167ecd1d06d1d7e",
+  "originHash" : "37844cf80fe1416dd9922e75c396bd616f0723f347f5649a1c0355a3ba3f3a82",
   "pins" : [
     {
       "identity" : "base45-swift",
