--- ios/Inji/RNEventEmitter.m
@@ -1,16 +1,0 @@
-#import <React/RCTBridgeModule.h>
-#import <React/RCTEventEmitter.h>
-
-@interface RCT_EXTERN_MODULE(RNEventEmitter, RCTEventEmitter)
-
-RCT_EXTERN_METHOD(emitEvent:(NSDictionary *)eventMap)
-
-RCT_EXTERN_METHOD(registerEventEmitter:(RCTPromiseResolveBlock)resolver
-                  rejecter:(RCTPromiseRejectBlock)rejecter)
-
-RCT_EXTERN_METHOD(emitMappedEvent:(id)event
-                  resolver:(RCTPromiseResolveBlock)resolver
-                  rejecter:(RCTPromiseRejectBlock)rejecter)
-
-
-@end
