+++ machines/bleShare/scan/scanServices.ts
@@ -16,7 +16,8 @@ import {
 } from '../../../shared/location';
 import {isIOS} from '../../../shared/constants';
 
-import {wallet, EventTypes, VerificationStatus} from '../../../tuvali';
+import {wallet, EventTypes, VerificationStatus} from '../../../shared/tuvali';
+import {WalletDataEvent} from '../../../shared/tuvali/types/events';
 
 export const ScanServices = (model: any) => {
   return {
