+++ shared/tuvali/index.tsx
@@ -1,6 +1,7 @@
 import {NativeEventEmitter, NativeModules, Platform} from 'react-native';
-import type {Verifier, Wallet} from './types/interface';
-import {isIOS} from '../constants';
+import type {Verifier, VersionModule, Wallet} from './types/interface';
+import {EventTypes, VerificationStatus} from './types/events';
+import {isAndroid, isIOS} from '../constants';
 
 const LINKING_ERROR =
   `The package 'tuvali' doesn't seem to be linked. Make sure: \n\n` +
@@ -9,9 +10,8 @@ const LINKING_ERROR =
   '- You are not using Expo Go\n';
 
 const VERIFIER_NOT_IMPLEMENTATED_ERROR = `Verifier is not yet implemented on IOS. Please remove Verifier usage on IOS Platform`;
-const isIOSPlatform = isIOS();
 
-export const wallet: Wallet = NativeModules.WalletModule
+const wallet: Wallet = NativeModules.WalletModule
   ? NativeModules.WalletModule
   : new Proxy(
       {},
@@ -22,35 +22,52 @@ export const wallet: Wallet = NativeModules.WalletModule
       },
     );
 
-export let verifier: Verifier = NativeModules.VerifierModule
-  ? NativeModules.VerifierModule
-  : new Proxy(
+// TODO: Use Actual Verifier module on IOS once Verifier is implemented
+let verifier: Verifier = NativeModules.VerifierModule;
+
+if (!isIOS()) {
+  if (verifier) setupModule(verifier);
+  else
+    new Proxy(
       {},
       {
         get() {
-          throw new Error(
-            isIOSPlatform ? VERIFIER_NOT_IMPLEMENTATED_ERROR : LINKING_ERROR,
-          );
+          throw new Error(VERIFIER_NOT_IMPLEMENTATED_ERROR);
         },
       },
     );
-
-if (!isIOSPlatform) {
-  setupModule(verifier);
 }
 
 setupModule(wallet);
 
+//
+// ErrorUtils.setGlobalHandler((error, isFatal) => {
+//   const eventEmitter = new NativeEventEmitter(NativeModules.Openid4vpBle);
+//
+//   console.error(
+//     `Exception in Tuvali: isFatal: ${isFatal}, error: ${JSON.stringify(
+//       error,
+//       null,
+//       4
+//     )}`
+//   );
+//
+//   eventEmitter.emit('EVENT_NEARBY', { type: 'onError', message: '' });
+//   Openid4vpBle.destroyConnection(() => {});
+// });
+
 function setupModule(module: any) {
-  if (Platform.OS === 'android') {
+  if (isAndroid()) {
     const eventEmitter = new NativeEventEmitter();
     module.handleDataEvents = (callback: (event: any) => void) =>
       eventEmitter.addListener('DATA_EVENT', callback);
   }
 
-  if (isIOSPlatform) {
+  if (isIOS()) {
     const eventEmitter = new NativeEventEmitter(module);
     module.handleDataEvents = (callback: (event: any) => void) =>
       eventEmitter.addListener('DATA_EVENT', callback);
   }
 }
+
+export {verifier, wallet, EventTypes, VerificationStatus};
