--- tuvali/tuvaliVersion.tsx
@@ -1,3 +1,0 @@
-const packageData = require('../package.json');
-
-export const tuvaliVersion = packageData.version.toString();
